Move some stuff around, change method name to be clearer, process only based on base type
