--- sample/CustomerApi/Services/CustomerInterests/Movie.cs
@@ -1,8 +1,0 @@
-using CustomerApi.Uris;
-
-namespace CustomerApi.Services.CustomerInterests;
-
-public record Movie(MovieUri MovieUri, string Name, int RuntimeMinutes) : CustomerInterest(MovieUri.MovieId, Name, StaticPartition)
-{
-    public const string StaticPartition = "/movies";
-}
