--- sample/CustomerApi/Services/CustomerInterests/Song.cs
@@ -1,8 +1,0 @@
-using CustomerApi.Uris;
-
-namespace CustomerApi.Services.CustomerInterests;
-
-public record Song(SongUri SongUri, string Name, string Genre) : CustomerInterest(SongUri.SongId, Name, StaticPartition)
-{
-    public const string StaticPartition = "/songs";
-}
